+++ examples/instance-segmenter.c
@@ -76,7 +76,7 @@ void train_isegmenter(char *datacfg, char *cfgfile, char *weightfile, int *gpus,
     args.d = &buffer;
     load_thread = load_data(args);
 
-    int epoch = (*net->seen)/N;
+    size_t epoch = (*net->seen)/N;
     while(get_current_batch(net) < net->max_batches || net->max_batches == 0){
         double time = what_time_is_it_now();
 
@@ -121,7 +121,7 @@ void train_isegmenter(char *datacfg, char *cfgfile, char *weightfile, int *gpus,
         if(*net->seen/N > epoch){
             epoch = *net->seen/N;
             char buff[256];
-            sprintf(buff, "%s/%s_%d.weights",backup_directory,base, epoch);
+            sprintf(buff, "%s/%s_%ld.weights",backup_directory,base, epoch);
             save_weights(net, buff);
         }
         if(get_current_batch(net)%100 == 0){
