+++ examples/tag.c
@@ -42,8 +42,8 @@ void train_tag(char *cfgfile, char *weightfile, int clear)
     fprintf(stderr, "%d classes\n", net->outputs);
 
     load_thread = load_data_in_thread(args);
-    int epoch = (*net->seen)/N;
-    while(get_current_batch(net) < net->max_batches || net->max_batches == 0){
+    size_t epoch = (*net->seen)/N;
+    while(get_current_batch(net) < (size_t) net->max_batches || net->max_batches == 0){
         time=clock();
         pthread_join(load_thread, 0);
         train = buffer;
@@ -59,7 +59,7 @@ void train_tag(char *cfgfile, char *weightfile, int clear)
         if(*net->seen/N > epoch){
             epoch = *net->seen/N;
             char buff[256];
-            sprintf(buff, "%s/%s_%d.weights",backup_directory,base, epoch);
+            sprintf(buff, "%s/%s_%ld.weights",backup_directory,base, epoch);
             save_weights(net, buff);
         }
         if(get_current_batch(net)%100 == 0){
